requirements.txt
aiogram==3.27.0
aiomysql==0.3.2
google-genai==1.73.1
mypy==1.20.0
PyMySQL==1.1.2
pytest==8.4.2
pytest-asyncio==0.26.0
python-dotenv==1.2.2
ruff==0.15.10
